Decouples ignite file from icp
wt will copy the ignite file into the worktree directory if it finds
one. This way, we don't need to rely on icp to exist or to contain the
ignite file.

We also deduplicate the logic into `do_icp` and do_ignite` functions
since they will both be invoked for add() and new() functions.